sentinel-core/pyproject.toml
[tool.poetry]
name = "sentinel-core"
version = "0.1.0"
description = "Core logic engine for Sentinel"
authors = ["Mystic"]
readme = "README.md"
packages = [{include = "sentinel_core"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.0"
sqlmodel = "^0.0.14"
ollama = "^0.1.0"
pypdf = "^6.6.2"
PyYAML = "^6.0.3"
tenacity = "^9.1.3"
httpx = "0.27.2"
jinja2 = "^3.1.6"
rich = "^13.0.0"
send2trash = "^1.8.0"
typer = "^0.9.0"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.6"
websockets = "^12.0"
pydantic-settings = "^2.0.0"

[tool.poetry.scripts]
sentinel = "sentinel_core.cli.main:app"


[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.6",
    "black>=23.11.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_backend"

# Ruff configuration
[tool.ruff]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "ARG", "PTH", "RUF"]
ignore = ["E501"]  # Line length handled by black

fixable = ["ALL"]
unfixable = []

exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

line-length = 100
target-version = "py311"

[tool.ruff.mccabe]
max-complexity = 10

# Black configuration
[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

# Mypy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
strict_equality = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=sentinel_core",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["sentinel_core"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
